Fix problem with background notification handling.
The default foreground state in the PushPlugin class was 'true'. Whilst
this would be true in the general case, if the app has been killed and
the GCMIntentService is started by the GCM BroadcastReceiver this should
in fact be false. So then, I'm switching the default value to 'false'.

onResume should be called when our main Activity is invoked, resetting
it to true, making all good.